+++ R/private_funs.R
@@ -1,5 +1,5 @@
-# Internal functions for librarian.
 
+# Internal functions for Librarian ---------------------------------------------
 
 #' Non-standard evaluation of dots with base R
 #'
@@ -136,6 +136,8 @@ check_attached <- function(packages = NULL) {
     }
 }
 
+
+
 #' Build a path, creating subfolders if needed
 #'
 #' Whereas `base::file.path()` only concatenates strings to build a path, `make_dirs()`
@@ -244,7 +246,107 @@ collapse_vec <- function(..., wrap = "'", collapse = ", ", unique = TRUE) {
 
 
 
-# Runs with devtools::release().
+#' Turn a list of words into a fuzzy regex
+#' 
+#' A fuzzy regex is one that will match search terms in any order by using PERL 
+#' lookaround. This is very slow, but often worth the cost to get more complete
+#' results.
+#'
+#' @param vec (Character) A string containing space-separated keywords to search for.
+#'
+#' @return A string where each word has been wrapped as a lookaround term.
+#'
+#' @examples
+#' \dontrun{
+#' fuzzy_needle("network centrality")
+#' #> [1] "(?=.*network)(?=.*centrality)"
+#' }
+fuzzy_needle <- function(vec) {
+    words <- unique(unlist(strsplit(vec, "\\s+")))
+    
+    groups <- sapply(words, function(x) paste0("(?=.*", x, ")"), USE.NAMES = FALSE)
+    
+    paste0(groups, collapse = "")
+}
+
+
+
+#' Truncate and wrap long text for console printing
+#'
+#' @param ... (Vectors) Vectors that will be joined together.
+#' @param trim (Int) The maximum length that `char` will be truncated to.
+#' @param width (Int) The target maximum column number to wrap the text at.
+#' @param indent (Int) Indent the first line by this many spaces.
+#' @param exdent (Int) Indent subsequent lines by this many spaces.
+#' @param prefix (Char) Begin every line with this string.
+#' @param initial (Char) Begin the first line with this string.
+#'
+#' @return A character vector.
+#'
+#' @examples
+#' \dontrun{
+#' wrap_long("This is a pretty long line that will need to be wrapped to 30 chars.",
+#'           trim = 50, width = 30) -> wrapped_text
+#' 
+#' cat(wrapped_text)
+#' 
+#' #>    This is a pretty long 
+#' #>    line that will need to 
+#' #>    be wr[...]
+#' }
+#' 
+#' @md
+wrap_long <- function(..., trim = 200, width = 80, indent = 4, exdent = indent, 
+                      prefix = "\n", initial = "") {
+    char <- c(...)
+    short_str <- strtrim(char, width = trim)
+    
+    if (nchar(short_str) < nchar(char)) {
+        short_str <- paste0(short_str, "[...]")
+    }
+    
+    strwrap(short_str, width = width, indent = indent, exdent = exdent, 
+            prefix = prefix, initial = initial)
+}
+
+
+
+is_valid_url <- function(string) {
+    # Automated testing fails with devtools::check() (but passes with devtools::test()) if
+    # the repo arg for install.packages() is not set properly. If I run getOption("repos")
+    # in R.exe running in the shell, I get the named vector c("CRAN" = "@CRAN@"), which is
+    # probably what was causing the error. To catch this, I'll test whether cran_repo is 
+    # a URL.
+    
+    # Regex is "@stephenhay" from https://mathiasbynens.be/demo/url-regex because it's the 
+    # shortest regex that matches every CRAN mirror at https://cran.r-project.org/mirrors.html
+    cran_repo_is_url <- grepl("(https?|ftp)://[^\\s/$.?#].[^\\s]*", cran_repo)
+}
+
+
+
+#' Assert that a URL is complete and valid
+#'
+#' @details The regex I use is "@stephenhay" from 
+#' <https://mathiasbynens.be/demo/url-regex> because it's the shortest regex that 
+#' matches every CRAN mirror at <https://cran.r-project.org/mirrors.html>.
+#'
+#' @param string (Character) A URL to check.
+#'
+#' @return A logical value, `TRUE` if the URL is valid, `FALSE` if otherwise.
+#'
+#' @examples
+#' is_valid_url("http://rstudio.com")
+#'
+#' @md
+is_valid_url <- function(string) {
+    any(grepl("(https?|ftp)://[^\\s/$.?#].[^\\s]*", string))
+}
+
+
+
+# Runs with devtools::release() ------------------------------------------------
+
 release_questions <- function() {
     c(
         "Have you run devtools::test()?"