+++ NEWS.md
@@ -3,6 +3,7 @@
 - MOD - `biocLite` has been replaced with `BiocManager` for installing BioConductor packages. `BiocManager` has been added as a dependency.
 - FIX - Various spelling errors in preparation for CRAN submission.
 - ADD - Vignette for CRAN
+- ADD - `browse_cran()` for searching CRAN package titles and descriptions by keyword or regular expression.